Add vm type field to ioctl args
We support vmlinux and bzimage guest images
for now, so add a type field in struct create_vm_args
and an enum that specifies which of the two we are creating.

Prior to the create_vm ioctl, parse the file data and set
the type field based on the result.

Signed-off-by: Connor Davis <davisc@ainfosec.com>