Fix sluggish guest by enabling MTRR/PAT
Linux only uses PAT if MTRRs are available. Not using PAT has a
*significant* performance impact on boot times (on the order of
minutes).

So emulate MTRRs so that Linux enables PAT.

An mtrr_range is created for each e820 entry given to the PVH domU.  The
range type is set to write-back. These ranges are used to emulate
reads/writes to the variable MTRRs.

The 'def type' MTRR is configured to disable fixed MTRRs, enable
variable MTRRs, and to map memory as UC. This means that any hole
in the e820 map will be mapped as UC.

Signed-off-by: Connor Davis <davisc@ainfosec.com>