[2/n] visr: Add visr class
To passthrough a device to a guest, there has to be coordination
between both the host and guest VM. The host VM needs to see
the device as a visr device so that it loads the visr driver
rather than the usual driver. The driver is responsible for reserving
1) an interrupt vector from the OS for later forwarding to the guest vcpu
and 2) BAR resources. The primary host-side APIs are:

- visr::emulate(bus, device, function)
    Emulate config space at the given b/d/f to present a device that
    matches with the driver in the host OS

- visr::enable()
    Install the required exit handlers to implement the emulation
    for devices added with previous emulate calls

The guest-side emulation is responsible for binding the virtual vector
the guest expects the device to use to a physical vector acquired
earlier by the host-side driver. This is done with:

- visr::bind_phys_vector(vcpuid)
    Return a physical vector received earlier from the host-side
    driver and associate it with the given vcpuid. Guest emulation
    should use this to program the actual device's MSI data register.

- visr::bind_virt_vector(vcpuid, virt)
    Associate the virtual vector number with the given vcpuid (and hence
    the physical vector associated previously through bind_phys_vector).
    Guest emulation is responsible for extracting this information from
    the guest so that visr can remap the physical interrupt to the
    vector the guest expects.