[5/n] visr: Add emulate ioctl and hypercall
Users can tell visr to emulate a PCI device by passing

    --emu bb:dd.f

to bfexec, where bb is the bus, dd is the device and f is the function
of the device (in hex). VISR will emulate the device's configuration
space and will present a virtual device that may be bound to the
kernel-level visr driver by rescanning the bus.

Signed-off-by: Connor Davis <davisc@ainfosec.com>